Move `useGetSupportInteractions` into the Help Center and disable for logged out users
